Create radosgw key on mon host
Instead of copying the bootstrap key to the rgw host, create the key on
the first mon host and just copy over the resulting key. The bootstrap
key basically gives full permissions to create keys on the cluster and
should be tightly secured and not be on a rgw host which might be
directly connected to the Internet.

If the pools are created by the Ansible playbook, the key permissions are
restricted to just allow the operations needed for radosgw and do not
allow pool manipulation.

This also changes the rgw_is_up test to use the admin key instead of the
bootstrap key because the bootstrap key no longer has sufficient
permissions to call "ceph -s".

Signed-off-by: Gaudenz Steinlin <gaudenz.steinlin@cloudscale.ch>