Fix RGW pool creation tests
The rgw_create_pools variable was populated with "fake" pools which did
not have the names expected by the RGW daemon. This only worked because
the RGW daemons created the needed pools on first start. The pools
created by rgw_create_pools were not used by the daemons at all. With
the change to restrict the permissions on the RGW key when the pools are
pre-created, this no longer worked.

Signed-off-by: Gaudenz Steinlin <gaudenz.steinlin@cloudscale.ch>